requirements.txt
pillow
rembg
aggdraw